refactorizacion de el directorio de libs
