+++ src/app/(Recipes pages)/functions/GetAllRecipes.ts
@@ -1,5 +1,5 @@
-import { baseUrl } from '@/lib/utils'
 import { type RecipeList } from '@interfaces/RecipesList.interface'
+import { baseUrl } from '@libs/utils'
 
 /**
  * @description Obtiene una lista de recetas desde el servidor con paginación.
@@ -9,14 +9,14 @@ import { type RecipeList } from '@interfaces/RecipesList.interface'
  */
 
 export async function GetAllRecipes(page = 1): Promise<RecipeList | undefined> {
-	try {
-		const response = await fetch(`${baseUrl}/recipes?page=${page}`)
-		if (!response.ok) {
-			throw new Error('Error al obtener todas las recetas')
-		}
-		const data: RecipeList = await response.json()
-		return data
-	} catch (error) {
-		console.error(error)
-	}
+  try {
+    const response = await fetch(`${baseUrl}/recipes?page=${page}`)
+    if (!response.ok) {
+      throw new Error('Error al obtener todas las recetas')
+    }
+    const data: RecipeList = await response.json()
+    return data
+  } catch (error) {
+    console.error(error)
+  }
 }
