+++ src/app/(Recipes pages)/functions/GetRecipeById.ts
@@ -1,5 +1,5 @@
-import { baseUrl } from '@/lib/utils'
 import { type Recipes } from '@interfaces/Recipes.interface'
+import { baseUrl } from '@libs/utils'
 
 /**
  * @description Obtiene una receta específica por su ID desde el servidor.
@@ -8,14 +8,14 @@ import { type Recipes } from '@interfaces/Recipes.interface'
  */
 
 export async function GetRecipeById(id: string) {
-	try {
-		const response = await fetch(`${baseUrl}/recipes/${id}`)
-		if (!response.ok) {
-			throw new Error('Error al obtener  la receta')
-		}
-		const data: Recipes = await response.json()
-		return data
-	} catch (error) {
-		console.error(error)
-	}
+  try {
+    const response = await fetch(`${baseUrl}/recipes/${id}`)
+    if (!response.ok) {
+      throw new Error('Error al obtener  la receta')
+    }
+    const data: Recipes = await response.json()
+    return data
+  } catch (error) {
+    console.error(error)
+  }
 }
