+++ src/app/(Recipes pages)/recetas/components/oneRecipe/ShowDataRecipe.tsx
@@ -3,7 +3,7 @@ import { GetRecipeById } from '@/(Recipes pages)/functions/GetRecipeById'
 import CheckBoxComponent from '@/(Recipes pages)/recetas/components/oneRecipe/CheckBox'
 import ChipComponent from '@/(Recipes pages)/recetas/components/oneRecipe/Chip'
 import { UiImage } from '@/components/ui'
-import GenerateJsonLD from '@/lib/GenerateJsonLD'
+import GenerateJsonLD from '@libs/GenerateJsonLD'
 import { notFound } from 'next/navigation'
 import { FiList } from 'react-icons/fi'
 import BreadcrumbsComponent from './Breadcrumbs'
