+++ src/libs/ResponsiveCaruselOptions.ts
